Auto-detect shell on Windows
